refactor: reconstruct package structure
