+++ src/main/java/kakao/controller/ReservationController.java
@@ -1,7 +1,7 @@
 package kakao.controller;
 
-import kakao.model.request.ReservationRequest;
-import kakao.model.response.ReservationResponse;
+import kakao.controller.request.ReservationRequest;
+import kakao.controller.response.ReservationResponse;
 import kakao.service.ReservationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
