+++ src/main/java/kakao/repository/JdbcTemplateReservationRepository.java
@@ -1,10 +1,10 @@
 package kakao.repository;
 
 import kakao.exception.ReservationNotFoundException;
-import kakao.model.entity.Reservation;
-import kakao.model.request.ReservationRequest;
-import kakao.model.response.ReservationResponse;
-import kakao.model.response.Theme;
+import kakao.model.Reservation;
+import kakao.controller.request.ReservationRequest;
+import kakao.controller.response.ReservationResponse;
+import kakao.model.Theme;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
